Finish to configure.
