+++ luna/__init__.py
@@ -1,9 +1,8 @@
 #coding=utf8
 
-from toolkit.config import (
-    parse_file,
-)
-from luna.core.exc import LunaException
-from luna.core.env import load_config
+"""
+Copyright 2015 Luna project
+"""
 
-config = load_config()
+
+from luna.core.config import config
