+++ luna/core/serve.py
@@ -4,13 +4,12 @@
 from __future__ import absolute_import, division, print_function
 
 import os
+
 from flask import Flask
-from luna import config
-from luna.core.api import Api
 from luna.core.logger import logger_init
 from luna.core.auth import auth_init
 from luna.core.db import db_init
-from luna.core.exc import exc_init
+from luna.core.config import config
 
 
 def api_init(app, apis):
@@ -25,21 +24,16 @@ def dynamic_load_apis():
         print(module)
 
 
-def init(settings):
+def init_app():
     app = Flask(__name__)
-
     app.config.update(config)
 
     apis = dynamic_load_apis()
     api_init(app, apis)
 
-    logger_init()
-
+    logger_init(app)
     auth_init(app)
-
-    db_init(settings.basic)
-
-    exc_init(app)
+    db_init(app)
 
     return app
 
@@ -59,7 +53,7 @@ def serve():
     app.run(host=app.config['HOST'], port=app.config['PORT'])
 
 
-app = init(settings)
+app = init_app()
 
 if __name__ == '__main__':
     serve()
